+++ app.py
@@ -1,34 +1,20 @@
 import streamlit as st
-import sqlalchemy
-from google.cloud.sql.connector import Connector, IPTypes
+from sqlalchemy import create_engine, select
+from sqlalchemy.orm import Session
+
+from data.setup import getconn
 
 ############# DATA FETCHING ##############
 
 # Connect with remote analytics DB
-
-# Initialize connector object
-connector = Connector()
-
-# Configure connector
-def getconn():
-
-    conn = connector.connect(
-        st.secrets["USER_ANALYTICS_DB_CONNECTION_NAME"],
-        "pymysql",
-        user=st.secrets["USER_ANALYTICS_DB_USERNAME"],
-        password=st.secrets["USER_ANALYTICS_DB_PASSWORD"],
-        db="user_analytics",
-        ip_type=IPTypes.PUBLIC,  # IPTypes.PRIVATE for private IP
-    )
-
-    return conn
-
-db_pool = sqlalchemy.create_engine(
+db_engine = create_engine(
     "mysql+pymysql://",
     creator=getconn,
 )
 
 # Fetch all records
+# with Session(db_engine) as db_session:
+    
 
 #
